requirements.txt
openai==1.12.0
psycopg2-binary==2.9.9
pgvector==0.2.4
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
matplotlib==3.8.2
python-dotenv==1.0.0
